Add new armhf substitution with filter-out on armhf in debian/rules, use it to pick right libGL.so
Upstream ships two (incompatible) versions of libGL, one is based on libglvnd and the other is not.
But on ARM only the latter is available. Add a new substitution in order to be able to pick libGL.so.VERSION
when building for ARM, and libGL.so.1.0.0 when building for amd64 and i386.


git-svn-id: svn://svn.debian.org/svn/pkg-nvidia/packages/nvidia-graphics-drivers/branches/361@6146 b6a4b1ba-b714-0410-8a2d-fc0b36c6496e